Update download.md
I added AnkerGames for preinstalled games it’s a really cool site! However, the only downside is that when a game link is dead, it just shows an error instead of indicating that the link is no longer active. I also added ReleaseBB for scene releases, though I’m unsure if it was a good choice since the downloads are quite slow. I’d love to hear your thoughts

Signed-off-by: Nikolay_Avilov <166333962+Nikolayavilov0@users.noreply.github.com>